pyproject.toml
[project]
name = "tds-project-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-sqlite3>=0.0.1",
    "duckdb>=1.2.0",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "markdown>=3.7",
    "openai>=1.63.0",
    "openai-function-calling>=2.3.0",
    "pandas>=2.2.3",
    "pybase64>=1.4.0",
    "pydantic>=2.10.6",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scipy>=1.15.1",
    "uvicorn>=0.34.0",
]
